Create CAR
Consider a car, and let’s call Hvθ(s) be the transfer function from the throttle angle command θ to the speed v. The throttle angle can vary between 0 and 43◦. The car is taken onto a test track and the engine is started. Starting from rest, the driver floors the gas pedal and holds it there. After a few seconds, the speed of the car reaches 120 mph and stays there. A table of speed vs time from the measurements during the test are assumed; where t = 0 corresponds to the time when the driver floors the gas pedal. Note that the throttle angle reaches its maximum value when the gas pedal is floored. The transfer function was estimated and a plot of predicted vs realistic speed is presented.